refactor: upgrade to Rust 2024 edition with comprehensive clippy improvements
- Upgrade Cargo.toml edition from 2021 to 2024
- Replace redundant closures with direct function references (4 instances)
- Modernize format strings to use inline format args (30+ instances)
- Add derivable Default implementations for DataCoding and ServiceType
- Simplify range validation with contains() methods
- Replace map_or patterns with is_ok_and() and direct comparisons
- Remove unnecessary type cast in deliver_sm.rs
- Add comprehensive safety documentation for unsafe function
- Maintain full backward compatibility and functionality

All 165 tests pass, code builds successfully, and clippy warnings
reduced from ~60 to 15 (mostly design suggestions).

+++ src/datatypes/deliver_sm.rs
@@ -377,7 +377,7 @@ impl DeliverSmBuilder {
     /// Build the DeliverSm, performing validation and calculating sm_length automatically
     pub fn build(mut self) -> Result<DeliverSm, DeliverSmValidationError> {
         // Auto-calculate sm_length from short_message
-        self.sm_length = self.short_message.len() as u8;
+        self.sm_length = self.short_message.len();
 
         let deliver_sm = DeliverSm {
             command_status: self.command_status,
@@ -669,9 +669,11 @@ mod tests {
 
         // Check that short message is included
         let message_bytes = "Hello World".as_bytes();
-        assert!(bytes
-            .windows(message_bytes.len())
-            .any(|window| window == message_bytes));
+        assert!(
+            bytes
+                .windows(message_bytes.len())
+                .any(|window| window == message_bytes)
+        );
     }
 
     #[test]
